+++ comics/comics/hipsterhitler.py
@@ -9,14 +9,15 @@ class Meta(MetaBase):
     rights = 'JC & APK'
 
 class Crawler(CrawlerBase):
-    history_capable_days = 40
+    history_capable_days = 120
     schedule = None
     time_zone = 0
 
     def crawl(self, pub_date):
         feed = self.parse_feed('http://hipsterhitler.com/feed/')
         for entry in feed.for_date(pub_date):
-            page = self.parse_page(entry.link)
-            url = page.src('img[src*="/wp-content/webcomic/"]')
+            if 'Comics' not in entry.tags:
+                continue
+            url = entry.content0.src('img')
             title = entry.title
             return CrawlerImage(url, title)