+++ comics/comics/lefthandedtoons.py
@@ -11,7 +11,7 @@ class Meta(MetaBase):
 
 class Crawler(CrawlerBase):
     history_capable_days = 12
-    schedule = 'Mo,Tu,We,Th,Fr'
+    schedule = 'Mo,Tu,We,Th'
     time_zone = -5
 
     def crawl(self, pub_date):